Move tools tests to src
